Rework tilemap engine

+++ src/Generic_Map.cpp
@@ -22,7 +22,7 @@ void Generic_Map::roll_left()
 {
     relative_position -= ROLL_STEP;
     
-    double factor = 1.f;
+    float factor = 1.f;
     for (int i = drawable_layers.size() - 1; i >= 1; --i)
     {
         drawable_layers[i].get_sprite().move({ -ROLL_STEP * factor, 0 });
@@ -34,7 +34,7 @@ void Generic_Map::roll_right()
 {
     relative_position += ROLL_STEP;
     
-    double factor = 1.f;
+    float factor = 1.f;
     for (int i = drawable_layers.size() - 1; i >= 1; --i)
     {
         drawable_layers[i].get_sprite().move({ ROLL_STEP * factor, 0 });
